Refactor integrations page components, fix some bugs
