+++ assets/src/components/integrations/IntegrationsOverview.tsx
@@ -1,313 +1,21 @@
 import React from 'react';
 import {RouteComponentProps} from 'react-router-dom';
-import dayjs from 'dayjs';
 import {Box, Flex} from 'theme-ui';
 import qs from 'query-string';
-import {
-  colors,
-  Button,
-  Divider,
-  Input,
-  Modal,
-  Paragraph,
-  Table,
-  Tag,
-  Text,
-  Title,
-} from '../common';
+import {Button, Paragraph, Text, Title} from '../common';
 import {PlusOutlined} from '../icons';
 import Spinner from '../Spinner';
-import {SLACK_CLIENT_ID} from '../../config';
 import * as API from '../../api';
-import {sleep} from '../../utils';
-
-type IntegrationType = {
-  key: 'slack' | 'gmail';
-  integration: string;
-  status: 'connected' | 'not_connected';
-  created_at?: string | null;
-  icon: string;
-};
-
-type WebhookEventSubscription = {
-  webhook_url: string;
-  verified: boolean;
-  created_at?: string | null;
-};
-
-const IntegrationsTable = ({
-  integrations,
-}: {
-  integrations: Array<IntegrationType>;
-}) => {
-  const columns = [
-    {
-      title: 'Integration',
-      dataIndex: 'integration',
-      key: 'integration',
-      render: (value: string, record: any) => {
-        const {icon} = record;
-
-        return (
-          <Flex sx={{alignItems: 'center'}}>
-            <img src={icon} alt={value} style={{height: 20}} />
-            <Text strong style={{marginLeft: 8}}>
-              {value}
-            </Text>
-          </Flex>
-        );
-      },
-    },
-    {
-      title: 'Status',
-      dataIndex: 'status',
-      key: 'status',
-      render: (value: string) => {
-        return value === 'connected' ? (
-          <Tag color={colors.green}>Connected</Tag>
-        ) : (
-          <Tag>Not connected</Tag>
-        );
-      },
-    },
-    {
-      title: 'Connected since',
-      dataIndex: 'created_at',
-      key: 'created_at',
-      render: (value: string) => {
-        if (!value) {
-          return '--';
-        }
-
-        return dayjs(value).format('MMMM DD, YYYY');
-      },
-    },
-    {
-      title: '',
-      dataIndex: 'action',
-      key: 'action',
-      render: (action: any, record: any) => {
-        const {key, status} = record;
-        const isConnected = status === 'connected';
-        // NB: when testing locally, update `origin` to an ngrok url
-        // pointing at localhost:4000 (or wherever your server is running)
-        const origin = window.location.origin;
-        const redirect = `${origin}/integrations/slack`;
-        const q = {
-          scope:
-            'incoming-webhook chat:write channels:history channels:manage chat:write.public users:read users:read.email',
-          user_scope: 'channels:history',
-          client_id: SLACK_CLIENT_ID,
-          redirect_uri: redirect,
-        };
-        const query = qs.stringify(q);
-
-        switch (key) {
-          case 'slack':
-            return (
-              <a href={`https://slack.com/oauth/v2/authorize?${query}`}>
-                <Button>{isConnected ? 'Reconnect' : 'Connect'}</Button>
-              </a>
-            );
-          default:
-            return <Button disabled>Coming soon!</Button>;
-        }
-      },
-    },
-  ];
-
-  return <Table dataSource={integrations} columns={columns} />;
-};
-
-const WebhooksTable = ({
-  webhooks,
-}: {
-  webhooks: Array<WebhookEventSubscription>;
-}) => {
-  const columns = [
-    {
-      title: 'Webhook URL',
-      dataIndex: 'webhook_url',
-      key: 'webhook_url',
-      render: (value: string, record: any) => {
-        return (
-          <Text keyboard strong>
-            {value}
-          </Text>
-        );
-      },
-    },
-    {
-      title: 'Status',
-      dataIndex: 'verified',
-      key: 'verified',
-      render: (verified: boolean) => {
-        return verified ? (
-          <Tag color={colors.green}>Verified</Tag>
-        ) : (
-          <Tag>Unverified</Tag>
-        );
-      },
-    },
-    {
-      title: 'Connected since',
-      dataIndex: 'created_at',
-      key: 'created_at',
-      render: (value: string) => {
-        if (!value) {
-          return '--';
-        }
-
-        return dayjs(value).format('MMMM DD, YYYY');
-      },
-    },
-    {
-      title: '',
-      dataIndex: 'action',
-      key: 'action',
-      render: (action: any, record: any) => {
-        return <Button>Update</Button>;
-      },
-    },
-  ];
-
-  return <Table rowKey="id" dataSource={webhooks} columns={columns} />;
-};
-
-// TODO: clean up a bit
-const NewWebhookModal = ({
-  visible,
-  onSuccess,
-  onCancel,
-}: {
-  visible: boolean;
-  onSuccess: (webhook: WebhookEventSubscription) => void;
-  onCancel: () => void;
-}) => {
-  const [url, setWebhookUrl] = React.useState('');
-  const [isVerifying, setIsVerifying] = React.useState(false);
-  const [isVerified, setIsVerified] = React.useState(false);
-  const [isSaving, setIsSaving] = React.useState(false);
-
-  const handleChangeUrl = (e: any) => setWebhookUrl(e.target.value);
-
-  const handleVerifyUrl = async () => {
-    console.log('Verifying:', url);
-    setIsVerifying(true);
-
-    const {verified} = await API.verifyWebhookUrl(url);
-    console.log('Verified?', verified);
-    await sleep(1000);
-
-    setIsVerifying(false);
-    setIsVerified(verified);
-  };
-
-  const handleCancelWebhook = () => {
-    onCancel();
-    setWebhookUrl('');
-    setIsVerified(false);
-  };
-
-  const handleSaveWebhook = async () => {
-    console.log('Saving:', url);
-    setIsSaving(true);
-
-    const webhook = await API.createEventSubscriptions({
-      webhook_url: url,
-    });
-
-    setIsSaving(false);
-    onSuccess(webhook);
-    setWebhookUrl('');
-    setIsVerified(false);
-  };
-
-  return (
-    <Modal
-      title="Add webhook URL"
-      visible={visible}
-      onOk={handleSaveWebhook}
-      onCancel={handleCancelWebhook}
-      footer={[
-        <Button key="cancel" onClick={handleCancelWebhook}>
-          Cancel
-        </Button>,
-        <Button
-          key="submit"
-          type="primary"
-          loading={isSaving}
-          onClick={handleSaveWebhook}
-        >
-          Save
-        </Button>,
-      ]}
-    >
-      <Box>
-        <Paragraph>
-          <Text>
-            You can subscribe to be notified of events in Papercups (for
-            example, when a new message is created) at a URL of your choice.
-          </Text>
-        </Paragraph>
-
-        <Box>
-          <Flex sx={{justifyContent: 'space-between', alignItems: 'center'}}>
-            <label htmlFor="webhook_url">
-              <Text strong>Webhook URL</Text>
-            </label>
-
-            {isVerifying ? (
-              <Text type="secondary">Verifying...</Text>
-            ) : isVerified ? (
-              <Text>Verified!</Text>
-            ) : null}
-          </Flex>
-          <Input
-            id="webhook_url"
-            size="large"
-            type="text"
-            value={url}
-            placeholder="https://myawesomeapp.com/api/webhook"
-            onChange={handleChangeUrl}
-            onBlur={handleVerifyUrl}
-          />
-        </Box>
-
-        <Divider />
-
-        <Paragraph>
-          <Text type="secondary">
-            We'll send HTTP POST requests to this URL when events occur. As soon
-            as you enter a URL, we'll send a request with a{' '}
-            <Text code>payload</Text> parameter, and your endpoint must respond
-            with the value.
-          </Text>
-        </Paragraph>
-
-        <Paragraph>
-          <Text type="secondary">
-            Note that for development URLs using <Text code>localhost</Text>,
-            you may need to use a tool like{' '}
-            <a
-              href="https://ngrok.com/"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              ngrok
-            </a>{' '}
-            to test your URL.
-          </Text>
-        </Paragraph>
-      </Box>
-    </Modal>
-  );
-};
+import {IntegrationType, WebhookEventSubscription} from './support';
+import IntegrationsTable from './IntegrationsTable';
+import WebhooksTable from './WebhooksTable';
+import NewWebhookModal from './NewWebhookModal';
 
 type Props = RouteComponentProps<{type?: string}> & {};
 type State = {
   loading: boolean;
   isWebhookModalOpen: boolean;
+  selectedWebhook: WebhookEventSubscription | null;
   integrations: Array<IntegrationType>;
   webhooks: Array<WebhookEventSubscription>;
 };
@@ -316,6 +24,7 @@ class IntegrationsOverview extends React.Component<Props, State> {
   state: State = {
     loading: true,
     isWebhookModalOpen: false,
+    selectedWebhook: null,
     integrations: [],
     webhooks: [],
   };
@@ -385,6 +94,10 @@ class IntegrationsOverview extends React.Component<Props, State> {
     this.setState({isWebhookModalOpen: true});
   };
 
+  handleUpdateWebhook = (webhook: WebhookEventSubscription) => {
+    this.setState({isWebhookModalOpen: true, selectedWebhook: webhook});
+  };
+
   refreshEventSubscriptions = async () => {
     try {
       const webhooks = await API.fetchEventSubscriptions();
@@ -397,21 +110,22 @@ class IntegrationsOverview extends React.Component<Props, State> {
 
   handleWebhookModalSuccess = (webhook: WebhookEventSubscription) => {
     this.setState({
-      webhooks: [...this.state.webhooks, webhook],
       isWebhookModalOpen: false,
+      selectedWebhook: null,
     });
 
     this.refreshEventSubscriptions();
   };
 
   handleWebhookModalCancel = () => {
-    this.setState({isWebhookModalOpen: false});
+    this.setState({isWebhookModalOpen: false, selectedWebhook: null});
   };
 
   render() {
     const {
       loading,
       isWebhookModalOpen,
+      selectedWebhook,
       webhooks = [],
       integrations = [],
     } = this.state;
@@ -469,11 +183,15 @@ class IntegrationsOverview extends React.Component<Props, State> {
           </Flex>
 
           <Box my={4}>
-            <WebhooksTable webhooks={webhooks} />
+            <WebhooksTable
+              webhooks={webhooks}
+              onUpdateWebhook={this.handleUpdateWebhook}
+            />
           </Box>
         </Box>
 
         <NewWebhookModal
+          webhook={selectedWebhook}
           visible={isWebhookModalOpen}
           onSuccess={this.handleWebhookModalSuccess}
           onCancel={this.handleWebhookModalCancel}
