+++ lib/chat_api/conversations.ex
@@ -84,7 +84,7 @@ defmodule ChatApi.Conversations do
 
   """
   def get_conversation!(id) do
-    Conversation |> Repo.get!(id) |> Repo.preload([:messages, :customer])
+    Conversation |> Repo.get!(id) |> Repo.preload([:customer, messages: [user: :profile]])
   end
 
   def get_conversation(id) do
